# backend/swaggerdocs.yaml
openapi: 3.0.0
info:
  title: Metatollapp API - OpenAPI 3.0
  description: |-
    MetatollappAPI
  termsOfService: http://swagger.io/terms/
  contact:
    name: Gkourgkoutas Spyridon
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:9103
...
paths:
  /interoperability/api:
    get:
      tags:
        - Admin
      description: Home API api
      responses:
        "200":
          description: "Success"
        "500":
          description: "Something went wrong"
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: X-OBSERVATORY-AUTH
      in: header